api-service/requirements.txt
fastapi
uvicorn[standard]
cvxpy
numpy
pandas
SQLAlchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic
pydantic-settings
google-genai
